requirements.txt
asgiref==3.7.2
Django==4.2.3
django-auth-ldap==4.4.0
djangorestframework==3.14.0
ldap3==2.9.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-ldap==3.4.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
PyJWT==2.7.0
